requirements.txt
Flask==3.1.3
bcrypt==5.0.0
PyJWT==2.12.1
Jinja2==3.1.6
Werkzeug==3.1.7
